# api/openapi.yaml
openapi: "3.0.3"
info:
  title: sahil-api
  description: |
    Live "now" data aggregator for sahil.im.
    Fetches current activity from Last.fm, AniList, Letterboxd, and Goodreads.
  version: "1.0.0"
  contact:
    name: Sahil A.
    url: https://sahil.im
servers:
  - url: https://api.sahil.im
    description: Production
  - url: http://localhost:8080
    description: Development
paths:
...
  /api/cms/feed:
    get:
      summary: Get feed entry
      operationId: getFeed
      parameters:
        - name: slug
          in: query
          schema:
            type: string
            default: feed
      responses:
        "200":
          description: Feed entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedEntry"
        "404":
          description: Not found
    post:
      summary: Upsert feed entry
      operationId: postFeed
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slug:
                  type: string
                  default: feed
                content:
                  type: string
              required:
                - content
      responses:
        "200":
          description: Feed entry saved
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    FeedEntry:
      type: object
      properties:
        slug:
          type: string
          example: "feed"
        title:
          type: string
          example: "My latest update"
        content:
          type: string
          example: "# Hello\n\nThis is my feed entry."
        updated:
          type: string
          format: date-time
          example: "2026-07-10T10:00:00Z"
...